refactored to move interface out to types
