+++ src/types/ModelSelectorProps.ts
@@ -1,0 +1,7 @@
+import { ModelInstance } from "./ModelInstance";
+
+export interface ModelSelectorProps {
+    href: string;
+    referenceId: string;
+    onModelSelect: (instance: ModelInstance) => void;
+  }
