+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './ModelInstance';
 export * from './ComplexLinkProps'; 
+export * from './ModelSelectorProps';
